+++ src/main/java/com/sprint/mission/discodeit/Repository/RepositoryService.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service;
+package com.sprint.mission.discodeit.Repository;
 
 import com.sprint.mission.discodeit.entity.BaseEntity;
