+++ src/main/java/com/sprint/mission/discodeit/Repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity.Repository;
+package com.sprint.mission.discodeit.Repository;
 
 import com.sprint.mission.discodeit.entity.Server;
