+++ src/main/java/com/sprint/mission/discodeit/service/jcf/JCFCategoryService.java
@@ -1,14 +1,9 @@
-package com.sprint.mission.discodeit.jcf;
+package com.sprint.mission.discodeit.service.jcf;
 
-import com.sprint.mission.discodeit.entity.Container.Container;
 import com.sprint.mission.discodeit.service.CategoryService;
-
-import java.util.LinkedList;
-import java.util.List;
 
 public class JCFCategoryService implements CategoryService {
     private static JCFCategoryService instance;
-    public List<Container> list;
 
     private JCFCategoryService() {
     }
@@ -18,10 +13,6 @@ public class JCFCategoryService implements CategoryService {
             instance = new JCFCategoryService();
         }
         return instance;
-    }
-
-    public void setList(LinkedList<Container> list) {
-        this.list = list;
     }
 
     @Override
@@ -34,7 +25,6 @@ public class JCFCategoryService implements CategoryService {
 
     }
 
-    //업데이트 조건, id가 동일해야한다.
     @Override
     public void update(String replaceName) {
 
@@ -42,12 +32,5 @@ public class JCFCategoryService implements CategoryService {
 
     @Override
     public void printCurrent() {
-        if (list.isEmpty()) {
-            System.out.println("카테고리와 채널이 없습니다.");
-            return;
-        }
-        for (int i = 0; i < list.size(); i++) {
-            System.out.println(i + 1 + " : " + list.get(i).getName());
-        }
     }
 }
