+++ src/main/java/com/sprint/mission/discodeit/service/jcf/JCFServerService.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.jcf;
+package com.sprint.mission.discodeit.service.jcf;
 
 import com.sprint.mission.discodeit.service.ServerService;
