+++ src/main/java/com/sprint/mission/discodeit/service/jcf/JCFUserService.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.jcf;
+package com.sprint.mission.discodeit.service.jcf;
 
 import com.sprint.mission.discodeit.Factory.CreateChannalFactory;
 import com.sprint.mission.discodeit.Factory.CreateServerFactory;
